# backend/openapi.yaml
...
openapi: 3.1.0
info:
  title: FantasyWiki API
  version: "1.0.0"
  summary: The HTTP contract of the FantasyWiki Worker.
  description: |
    A fantasy league played with Wikipedia articles. Players found leagues, buy
    contracts on articles out of a Top Read Snapshot, field a formation, and are
    scored on how much those articles were read.

    Three authentication regimes sit side by side, and which one applies is
    decided by the path prefix rather than by the operation:

    | Prefix | Who calls it | How |
    |---|---|---|
    | `/auth/*` | A browser being signed in | Nothing yet — this is where a session is minted |
    | `/api/v1/*` | The frontend, as a signed-in player | The `session_token` cookie, HTTP-only |
    | `/internal/v1/*` | The scoring engine | A bearer service token |

    The major version lives in the path segment (`v1`), not in a header or a
    query parameter, so a future `v2` can be mounted beside `v1` while clients
    migrate; `version` above is the contract version, and its major component
    must always equal that segment.

    Identity is always resolved server-side, from the session or the token. No
    endpoint takes a `playerId` from the client, which is why the self-scoped
    reads are spelled `/api/v1/me` and `my-…` rather than carrying an id in the
    path.

    Every failure answers the same shape, a single `error` field carrying a
    constant the frontend matches on. The constant is never parsed for meaning;
...
  license:
...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /internal/v1/scoring-inputs:
    get:
      operationId: getScoringInputs
      tags: [Internal]
      summary: The day's scorable teams
      description: >
        Every team with a line-up to score on the given day, flattened into the
        shape the scoring engine consumes: the articles it fielded and the
        Chemistry Link topology of its schema.
      security:
        - serviceToken: []
      parameters:
        - name: date
          in: query
          required: true
          description: The day to score. An impossible date is rejected rather than rolled over.
          schema:
            type: string
            format: date
          example: "2026-08-26"
      responses:
        "200":
          description: One entry per scorable team. Empty when nothing was fielded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoringInput"
        "400":
          description: The date parameter is missing or not a real date.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /internal/v1/performances:
    post:
      operationId: ingestPerformances
      tags: [Internal]
      summary: Write back the day's computed performances
      description: >
        Idempotent and chunkable: the engine may send a day in several calls and
        may resend one it already sent, which is what lets a failed run be
        retried without reconciling first.
      security:
        - serviceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerformanceIngest"
      responses:
        "200":
          description: How many rows the call wrote.
          content:
            application/json:
              schema:
                type: object
                required: [written]
                properties:
                  written:
                    type: integer
                    minimum: 0
        "400":
          description: Malformed JSON, a missing or impossible date, or a result the service refused.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/session:
    get:
      operationId: getSession
      tags: [Session]
      summary: The signed-in player
      description: >
        Read straight off the JWT rather than out of the database — these are
        the claims the cookie carries, not a player record. `features` is read
        off the Worker's own bindings, so a deployment that was never given the
        Workers AI model reports the Genie as absent whatever any config says.
      responses:
        "200":
          description: The session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: signOut
      tags: [Session]
      summary: Sign out
      description: >
        Expires the session cookie. Nothing server-side is revoked — the JWT is
        stateless, and a copy taken before this call stays valid until it
        expires.
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_token
      description: >
        A seven-day HS256 JWT, set by the sign-in redirect and HTTP-only, so no
        script — the frontend's own included — can read it. The frontend simply
        sends every request with credentials included. Identity is the `sub`
        claim; nothing about who is asking is ever taken from a URL or a body.
    serviceToken:
      type: http
      scheme: bearer
      description: >
        The shared secret the scoring engine presents on /internal/v1/*. Compared
        in constant time, and an unset secret fails closed rather than open.
...
  responses:
    Success:
      description: Done.
      content:
        application/json:
          schema:
            type: object
            required: [success]
            properties:
              success:
                type: boolean
                const: true
    Unauthorized:
      description: No session cookie, or one that has expired or does not verify.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ServerError:
      description: >
        Something nobody named — a D1 outage, an upstream failure. Never a
        refusal the caller could act on: those are 4xx and carry a constant.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: >
        Every failure, in one shape. `error` carries a constant the frontend
        matches on exactly; its wording is not part of the contract and must
        never be parsed for meaning.
      required: [error]
      properties:
        error:
          type: string
          example: NOT_ENOUGH_CREDITS
...
    Session:
      type: object
      required: [sub, email, name, picture, features]
      properties:
        sub:
          type: string
          description: The Google account id the session was minted for.
        email:
          type: string
          format: email
        name:
          type: string
        picture:
          type: string
          description: Avatar URL. Empty for the demo player.
        features:
          type: object
          description: What this deployment can offer, read off its own bindings.
          required: [articleGenie]
          properties:
            articleGenie:
              type: boolean
...
    Domain:
      type: string
      description: >
        A Wikipedia language edition code, as Wikimedia writes them — lowercase
        letters, digits and hyphens. What an edition may actually host a league
        on is decided against live data at league creation, not by this shape.
      maxLength: 20
      pattern: "^[a-z][a-z0-9]*(-[a-z0-9]+)*$"
      examples: ["en", "it", "pt-br"]
...
    ChemistryLevel:
      type: string
      description: >
        How strongly two adjacent articles link to each other. Mutual links are
        excellent, one-way good, neither weak; empty is an unfilled slot.
      enum: [empty, weak, good, excellent]
...
    ScoringInput:
      type: object
      description: One scorable team, flattened into what the scoring engine needs and nothing else.
      required:
        [leagueId, teamId, domain, articles, chemistryLinks, formationSnapshot]
      properties:
        leagueId:
          type: string
        teamId:
          type: string
        domain:
          $ref: "#/components/schemas/Domain"
        articles:
          type: array
          description: The canonical titles fielded that day.
          items:
            type: string
        chemistryLinks:
          type: array
          description: The schema's topology as article pairs, already resolved from positions.
          items:
            type: array
            minItems: 2
            maxItems: 2
            items:
              type: string
        formationSnapshot:
          type: string
          description: >
            The position-to-article map, pre-serialised. Opaque to the backend,
            which stores it verbatim and hands it back untouched.
    PerformanceResult:
      type: object
      required: [teamId, articleViews, chemistryLevels, formationSnapshot]
      properties:
        teamId:
          type: string
        articleViews:
          type: array
          items:
            type: integer
        chemistryLevels:
          type: array
          items:
            $ref: "#/components/schemas/ChemistryLevel"
        formationSnapshot:
          type: string
          description: Echoed back from the scoring input it was computed from.
    PerformanceIngest:
      type: object
      required: [date, results]
      properties:
        date:
          type: string
          format: date
          example: "2026-08-26"
        results:
          type: array
          items:
            $ref: "#/components/schemas/PerformanceResult"