# backend/src/routes/me.ts
import { Hono } from "hono";
import {
  GENIE_ERRORS,
  GenieSeedRequest,
  GenieTurnRequest,
} from "../../../dto/genieDTO";
import { ArticleGenieService } from "../services/articleGenie";
import { WorkersAiBinding, createLlmClient } from "../services/llmClient";
import { AppVariables } from "../appEnv";
import { currentPlayer } from "./currentPlayer";

interface RateLimiter {
  limit(options: { key: string }): Promise<{ success: boolean }>;
}

type Bindings = {
  /**
   * Optional because the `local` environment deliberately leaves the Workers AI
   * binding out (see `wrangler.jsonc`): it has no local simulator, so declaring
   * it would make `wrangler dev` demand Cloudflare credentials that a fresh
   * clone has none of, and the whole Worker would fail to start over one
   * optional feature. Where it is absent the Genie is simply asleep.
   */
  AI?: WorkersAiBinding;
  GENIE_RATE_LIMITER: RateLimiter;
};

/**
 * Self-scoped routes for the authenticated player: identity comes from the JWT
 * and no `playerId` is ever accepted from the client
 * (docs/development/api-naming-rules.md).
 */
const me = new Hono<{ Bindings: Bindings; Variables: AppVariables }>();

/**
 * Everything the client got wrong about its own request. The Genie is additive,
 * so none of these should ever reach a player — they mean the frontend built a
 * payload it shouldn't have, and a 400 says so instead of burning a model call.
 */
const VALIDATION_ERRORS: string[] = [
  GENIE_ERRORS.MALFORMED_BODY,
  GENIE_ERRORS.QUERY_REQUIRED,
  GENIE_ERRORS.QUERY_TOO_LONG,
  GENIE_ERRORS.TOO_MANY_CANDIDATES,
  GENIE_ERRORS.NO_CANDIDATES,
  GENIE_ERRORS.HISTORY_TOO_LONG,
  GENIE_ERRORS.TEXT_TOO_LONG,
  GENIE_ERRORS.INVALID_BUCKET,
  GENIE_ERRORS.DUPLICATE_CANDIDATE_ID,
];

/**
 * A deployment without the Workers AI binding has no genie to wake. Checked
 * before the rate limiter so a build that never offers the feature cannot spend
 * the player's budget on it, and answered with `ASLEEP` because that is what
 * the absence means to a player — the same code quota exhaustion already
 * returns, and the same one the frontend already knows how to dismiss.
 *
 * Belt and braces: `GET /api/v1/session` reports the same fact as a feature flag
 * and the frontend hides the Genie's entry point, so this should be unreachable
 * from the app itself.
 */
function llmFor(ai: WorkersAiBinding | undefined) {
  return ai === undefined ? undefined : createLlmClient(ai);
}

/**
 * Reads the player's opening text into the searches that seed the candidate
 * set, so the browser can go and run them.
 *
 * Separate from a turn only because of ordering: a turn is defined over a
 * candidate list, and there is no list until the query has been read. Seeding a
 * chemistry query on keywords alone would have the Genie open with a question
 * about articles *describing* OpenAI rather than ones relating it to Portugal.
 */
me.post("/genie-seeds", currentPlayer, async (c) => {
  const llm = llmFor(c.env.AI);
  if (llm === undefined) {
    return c.json({ error: GENIE_ERRORS.ASLEEP }, 503);
  }

  const { success } = await c.env.GENIE_RATE_LIMITER.limit({
    key: c.var.player.id,
  });
  if (!success) {
    return c.json({ error: GENIE_ERRORS.RATE_LIMITED }, 429);
  }

  let request: GenieSeedRequest;
  try {
    request = await c.req.json<GenieSeedRequest>();
  } catch {
    return c.json({ error: GENIE_ERRORS.MALFORMED_BODY }, 400);
  }

  const service = new ArticleGenieService(llm);
  const result = await service.seed(request);

  if (!result.ok) {
    const status = VALIDATION_ERRORS.includes(result.error) ? 400 : 503;
    return c.json({ error: result.error }, status);
  }

  return c.json(result.value);
});

/**
 * One turn of the Article Genie (ADR 0006). The Worker does exactly one thing
 * here — call the model — because every other part of the feature is cheaper in
 * the browser: the Wikimedia search and link fetches run against the client's
 * 7-day cache, where neither the 50-subrequest nor the 10 ms CPU limit applies.
 *
 * A session is several requests in quick succession, so the rate limiter is
 * sized for real play rather than for abuse; the day's neuron allocation is the
 * real ceiling, and it fails closed on its own.
 */
me.post("/genie-turns", currentPlayer, async (c) => {
  const llm = llmFor(c.env.AI);
  if (llm === undefined) {
    return c.json({ error: GENIE_ERRORS.ASLEEP }, 503);
  }

  const { success } = await c.env.GENIE_RATE_LIMITER.limit({
    key: c.var.player.id,
  });
  if (!success) {
    return c.json({ error: GENIE_ERRORS.RATE_LIMITED }, 429);
  }

  let request: GenieTurnRequest;
  try {
    request = await c.req.json<GenieTurnRequest>();
  } catch {
    return c.json({ error: GENIE_ERRORS.MALFORMED_BODY }, 400);
  }

  const service = new ArticleGenieService(llm);
  const result = await service.takeTurn(request);

  if (!result.ok) {
    // Anything that is not the client's fault is the genie being asleep, which
    // the frontend answers by dismissing to the ordinary search bar. 503 rather
    // than 502: the day's allocation coming back is a matter of waiting.
    const status = VALIDATION_ERRORS.includes(result.error) ? 400 : 503;
    return c.json({ error: result.error }, status);
  }

  return c.json(result.value);
});

export default me;